Create CRLF line endings
Line ending Correction in case the script has been created on a non Linux System Such as Windows System which uses CRLF (Carraige Return and Line Feed) Instead of the Unix/Linux LF line endings. The ^M character represents the carriage return (\r) that Linux doesn't recognize. Therefore the script helps to change the line ending in order for the main Script to be recorgnized and be successfully executed on the Linux System.